+++ Widgets/QuestWidget.lua
@@ -4,7 +4,7 @@
 local ADDON_NAME, Addon = ...
 local ThreatPlates = Addon.ThreatPlates
 
-local Widget = (Addon.IS_MAINLINE and Addon.Widgets:NewWidget("Quest")) or {}
+local Widget = (Addon.ExpansionIsAtLeastMists and Addon.Widgets:NewWidget("Quest")) or {}
 
 ---------------------------------------------------------------------------------------------------
 -- Imported functions and constants
@@ -20,14 +20,16 @@ local UnitExists = UnitExists
 local IsInRaid, IsInGroup, GetNumGroupMembers, GetNumSubgroupMembers = IsInRaid, IsInGroup, GetNumGroupMembers, GetNumSubgroupMembers
 local wipe = wipe
 
-local RequestLoadQuestByID = C_QuestLog.RequestLoadQuestByID
 local GetQuestObjectives, GetQuestInfo = C_QuestLog.GetQuestObjectives, C_QuestLog.GetInfo
-local GetLogIndexForQuestID, GetNumQuestLogEntries = C_QuestLog.GetLogIndexForQuestID, C_QuestLog.GetNumQuestLogEntries
+local RequestLoadQuestByID = C_QuestLog.RequestLoadQuestByID
+local GetLogIndexForQuestID = C_QuestLog and C_QuestLog.GetLogIndexForQuestID or GetQuestLogIndexByID
+local GetNumQuestLogEntries = C_QuestLog and C_QuestLog.GetNumQuestLogEntries or GetNumQuestLogEntries
+local C_TooltipInfo_GetUnit = C_TooltipInfo and C_TooltipInfo.GetUnit -- Added in 10.0.2
 local GetNamePlateForUnit = C_NamePlate.GetNamePlateForUnit
-local C_TooltipInfo_GetUnit = C_TooltipInfo and C_TooltipInfo.GetUnit
 
 -- ThreatPlates APIs
 local PlayerName = Addon.PlayerName
+local RGB_P = ThreatPlates.RGB_P
 local UnitDetailedThreatSituationWrapper = Addon.UnitDetailedThreatSituationWrapper
 
 local _G =_G
@@ -56,51 +58,104 @@ local GroupMembers = {}
 
 local IsQuestUnit -- Function
 
+if Addon.ExpansionIsBetween(LE_EXPANSION_MISTS_OF_PANDARIA, LE_EXPANSION_DRAGONFLIGHT) then
+  local ScannerName = "ThreatPlates_Tooltip_Subtext"
+  local TooltipScanner = CreateFrame( "GameTooltip", ScannerName , nil, "GameTooltipTemplate" ) -- Tooltip name cannot be nil
+  TooltipScanner:SetOwner( WorldFrame, "ANCHOR_NONE" )
+
+  local TooltipScannerData = {
+    lines = {
+      [1] = {},
+      [2] = {},
+      [3] = {},
+      [4] = {},
+      [5] = {},
+    }
+  }
+
+  local function CreateLineData(line)
+    local line_id = ScannerName .. "TextLeft" .. tostring(line)
+
+    TooltipScannerData.lines[line].leftText = _G[line_id]:GetText()
+    TooltipScannerData.lines[line].leftColor = RGB_P(_G[line_id]:GetTextColor())
+  end
+
+  -- Compatibility functions for tooltips in WoW Classic
+  C_TooltipInfo_GetUnit = function(unitid)
+    TooltipScanner:ClearLines()
+		TooltipScanner:SetUnit(unitid)
+
+    CreateLineData(3)
+    CreateLineData(4)
+    CreateLineData(5)
+    
+    return TooltipScannerData
+  end
+end
+
+ -- C_TooltipInfo.GetUnit: Added in 10.0.2
+if not GetQuestInfo then
+  GetQuestInfo = function(quest_log_index)
+    --local title, level, suggestedGroup, isHeader, isCollapsed, isComplete, frequency, questID, startEvent, displayQuestID, isOnMap, hasLocalPOI, isTask, isBounty, isStory, isHidden, isScaling
+    local title, _, _, isHeader, _, _, _, questID = GetQuestLogTitle(quest_log_index)    
+    return {
+        title = title,
+        isHeader = isHeader,
+        questID = questID,        
+    }
+  end
+end
+
 -- Since patch 8.3, quest tooltips have a different format depending on the localization, it seems
 -- at least for kill quests
 -- In Shadowlands, it seems that the format is randomly changed, at least for German, so check for
 -- everything as a backup
-local QUEST_OBJECTIVE_PARSER_LEFT = function(text)
-  local current, goal, objective_name = string.match(text,"^(%d+)/(%d+)( .*)$")
-
-  if not objective_name then
-    objective_name, current, goal = string.match(text,"^(.*: )(%d+)/(%d+)$")
-  end
-
-  return objective_name, current, goal
-end
-
-local QUEST_OBJECTIVE_PARSER_RIGHT = function(text)
-  -- Quest objective: Versucht, zu kommunizieren: 0/1
-  local objective_name, current, goal = string.match(text,"^(.*: )(%d+)/(%d+)$")
-
-  if not objective_name then
-    -- Quest objective: 0/1 Besucht die Halle der Kuriositäten
-    current, goal, objective_name = string.match(text,"^(%d+)/(%d+)( .*)$")
-  end
-
-  return objective_name, current, goal
-end
-
-local STANDARD_QUEST_OBJECTIVE_PARSER = {
-  -- x/y Objective
-  enUS = QUEST_OBJECTIVE_PARSER_LEFT,
-  -- enGB = enGB clients return enUS
-  esMX = QUEST_OBJECTIVE_PARSER_LEFT,
-  ptBR = QUEST_OBJECTIVE_PARSER_LEFT,
-  itIT = QUEST_OBJECTIVE_PARSER_LEFT,
-  koKR = QUEST_OBJECTIVE_PARSER_LEFT,
-  zhTW = QUEST_OBJECTIVE_PARSER_LEFT,
-  zhCN = QUEST_OBJECTIVE_PARSER_LEFT,
-
-  -- Objective: x/y
-  deDE = QUEST_OBJECTIVE_PARSER_RIGHT,
-  frFR = QUEST_OBJECTIVE_PARSER_RIGHT,
-  esES = QUEST_OBJECTIVE_PARSER_RIGHT,
-  ruRU = QUEST_OBJECTIVE_PARSER_RIGHT,
+local OBJECTIVE_PARSER_REGEXP_BY_EXPANSION = {
+  [LE_EXPANSION_MISTS_OF_PANDARIA] = {
+    -- Format (Mists): " - Objective: x/y"
+    LEFT =  "^%s.%s(.*: )(%d+)/(%d+)$",
+    RIGHT = "^%s.%s(.*: )(%d+)/(%d+)$",
+  },
+  MAINLINE = {
+    -- Format: "x/y Objective"
+    LEFT = "^(%d+)/(%d+)( .*)$",
+    -- Format: "Objective: x/y"
+    RIGHT = "^(.*: )(%d+)/(%d+)$",
+  },
 }
 
-local QuestObjectiveParser = STANDARD_QUEST_OBJECTIVE_PARSER[GetLocale()] or QUEST_OBJECTIVE_PARSER_LEFT
+local OBJECTIVE_GOAL_ALIGNMENT = {
+  enUS = "LEFT",
+  -- enGB = enGB clients return enUS
+  esMX = "LEFT",
+  ptBR = "LEFT",
+  itIT = "LEFT",
+  koKR = "LEFT",
+  zhTW = "LEFT",
+  zhCN = "LEFT",
+  --
+  deDE = "RIGHT",
+  frFR = "RIGHT",
+  esES = "RIGHT",
+  ruRU = "RIGHT",
+}
+
+-- Set correct regexps for current expansion and locale
+local RegexpForExpansion = OBJECTIVE_PARSER_REGEXP_BY_EXPANSION[Addon.GetExpansionLevel()]
+local RegexpAlignment = OBJECTIVE_GOAL_ALIGNMENT[GetLocale()]
+
+local RegexpDefault = (RegexpAlignment == "LEFT" and RegexpForExpansion.LEFT) or RegexpForExpansion.RIGHT
+local RegexpBackup = (RegexpAlignment == "LEFT" and RegexpForExpansion.RIGHT) or RegexpForExpansion.LEFT
+
+local function QuestObjectiveParser(text)
+  local objective_name, current, goal = string.match(text, RegexpDefault)
+
+  if not objective_name then
+    current, goal, objective_name = string.match(text, RegexpBackup)
+  end
+
+  return objective_name, current, goal
+end
 
 ---------------------------------------------------------------------------------------------------
 -- Quest Functions
@@ -116,9 +171,11 @@ function IsQuestUnit(unit)
 
   for i = 3, #tooltip_data.lines do
     local line = tooltip_data.lines[i]
-
     local text = line.leftText
     local text_r, text_g, text_b = line.leftColor.r, line.leftColor.g, line.leftColor.b
+
+    -- At list in Mists, tooltips are not guarantted to have 5 lines min.
+    if not text then break end
 
     if text_r > 0.99 and text_g > 0.8 and text_b == 0 then
       -- A line with this color is either the quest title or a player name (if on a group quest, but always after the quest title)
@@ -215,7 +272,7 @@ local function CacheQuestObjectives(quest)
   quest.Objectives = quest.Objectives or {}
 
   local all_objectives = GetQuestObjectives(quest.questID)
-
+  
   local objective
   for objIndex = 1, #all_objectives do
     objective = all_objectives[objIndex]
@@ -319,7 +376,7 @@ function Widget:QUEST_LOG_UPDATE()
   -- UnitQuestLogChanged being true means that UNIT_QUEST_LOG_CHANGED was fired (possibly several times)
   -- So there should be quest progress => update all plates with the current progress.
   if UnitQuestLogChanged then
-    --print ("QUEST_LOG_UPDATE => UnitQuestLogChanged")
+    --f ("QUEST_LOG_UPDATE => UnitQuestLogChanged")
     UnitQuestLogChanged = false
 
     -- Update the cached quest progress (for non-progressbar quests) after QUEST_WATCH_UPDATE
@@ -338,11 +395,12 @@ function Widget:QUEST_LOG_UPDATE()
     -- We need to do this to update all progressbar quests - their quest progress cannot be cached
     self:UpdateAllFramesAndNameplateColor()
   end
-
+  
   -- It does seem that this is no longer necessary
   if QuestLogNotComplete then
     QuestLogNotComplete = false
     self:GenerateQuestCache()
+    self:UpdateAllFramesAndNameplateColor()
   end
 end