+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 # @project-version@ (@build-time@)
 
-* Fixed a bug that was preventing the display of Chi points for Monks in the Combo Points widget.
+* Enabled Quest widget for Mists Classic as quest tooltips are now available there.